Create $disabledaccounts
This is to pull disabled accounts in AD, select only the samaccountname, enabled, and lastlogondate object and then sort them by lastlogondate. Then it will export them to a csv. If there are no disabled accounts though it will tell you there are none.

This is my first if else so if anyone has any feed back i would appreciate it.